requirements.txt
Flask
python-dotenv==1.0.0
pandas
pandasai
langchain_groq
langchain_openai
langchain_mistralai
cryptography
numpy==1.26.4
